add auto post, swear, email

+++ breadcumb_search.php
@@ -55,21 +55,30 @@ $json = json_decode($string, true);
 						require_once ('incl/server.php');
 						$sql="SELECT * FROM skelbimai WHERE id='$item'";
 						$result=sqlconnect($sql);
-						while ($row = $result->fetch_assoc()) {
-							$cat1=$row['cat1'];
-							$cat2=$row['cat2'];
-							$make=$row['make'];
-							$model=$row['model'];
-							
+						$ad_count=$result->num_rows;
+						if($ad_count>0){
+							while ($row = $result->fetch_assoc()) {
+								$cat1=$row['cat1'];
+								$cat2=$row['cat2'];
+								$make=$row['make'];
+								$model=$row['model'];
+								
+							}
+						}else{
+								$cat1='';
+								$cat2='';
+								$make='';
+								$model='';
 						}
 						
 						$link='items/'.str_replace(' ','-',$cat1);
 						echo '<li><i class="fa fa-angle-right" aria-hidden="true"></i></li>';
 						echo '<li><a href="/'.$link.'">'.str_replace("-"," ",$cat1).'</a></li>';
 						
-						$link.='/'.str_replace(' ','-',$cat2);
-						echo '<li><i class="fa fa-angle-right" aria-hidden="true"></i></li>';
-						echo '<li><a href="/'.$link.'">'.str_replace("-"," ",$cat2).'</a></li>';
+						if($cat2!==''){$link.='/'.str_replace(' ','-',$cat2);
+							echo '<li><i class="fa fa-angle-right" aria-hidden="true"></i></li>';
+							echo '<li><a href="/'.$link.'">'.str_replace("-"," ",$cat2).'</a></li>';
+						}
 						
 						if($make!==''){$link.='/'.str_replace(' ','-',$make);
 							echo '<li><i class="fa fa-angle-right" aria-hidden="true"></i></li>';
